requirements.txt
Django
openpyxl
pandas
python-docx
pdfplumber
PyPDF2